requirements.txt
geocoder
geopy
google-genai
numpy
opencv-python-headless
streamlit
firebase-admin